+++ src/test/java/io/jenkins/plugins/opentelemetry/JenkinsOtelPluginIntegrationTest.java
@@ -6,94 +6,35 @@
 package io.jenkins.plugins.opentelemetry;
 
 import com.github.rutledgepaulv.prune.Tree;
-import hudson.ExtensionList;
 import hudson.Functions;
 import hudson.model.FreeStyleBuild;
 import hudson.model.FreeStyleProject;
 import hudson.model.Node;
 import hudson.model.Result;
-import io.jenkins.plugins.casc.misc.ConfiguredWithCode;
-import io.jenkins.plugins.casc.misc.JenkinsConfiguredWithCodeRule;
 import io.jenkins.plugins.opentelemetry.semconv.JenkinsOtelSemanticAttributes;
 import io.jenkins.plugins.opentelemetry.semconv.JenkinsSemanticMetrics;
-import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.metrics.data.LongPointData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
 import io.opentelemetry.sdk.metrics.data.MetricDataType;
 import io.opentelemetry.sdk.testing.exporter.InMemoryMetricExporter;
-import io.opentelemetry.sdk.testing.exporter.InMemorySpanExporter;
-import io.opentelemetry.sdk.trace.data.SpanData;
 import org.hamcrest.CoreMatchers;
 import org.hamcrest.MatcherAssert;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.job.WorkflowJob;
 import org.jenkinsci.plugins.workflow.job.WorkflowRun;
-import org.junit.*;
-import org.jvnet.hudson.test.BuildWatcher;
+import org.junit.Test;
 
-import java.util.*;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.BiPredicate;
-import java.util.logging.Logger;
-import java.util.stream.Collectors;
-
-import static com.google.common.base.Verify.verify;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 /**
  * Note usage of `def xsh(cmd) {if (isUnix()) {sh cmd} else {bat cmd}}` is inspired by
  * https://github.com/jenkinsci/workflow-basic-steps-plugin/blob/474cea2a53753e1fb9b166fa1ca0f6184b5cee4a/src/test/java/org/jenkinsci/plugins/workflow/steps/IsUnixStepTest.java#L39
  */
-public class JenkinsOtelPluginIntegrationTest {
-    static {
-        OpenTelemetrySdkProvider.TESTING_INMEMORY_MODE = true;
-        OpenTelemetrySdkProvider.TESTING_SPAN_EXPORTER = InMemorySpanExporter.create();
-        OpenTelemetrySdkProvider.TESTING_METRICS_EXPORTER = InMemoryMetricExporter.create();
-    }
-
-    private final static Logger LOGGER = Logger.getLogger(JenkinsOtelPluginIntegrationTest.class.getName());
-
-    final static AtomicInteger jobNameSuffix = new AtomicInteger();
-
-    @ClassRule
-    public static BuildWatcher buildWatcher = new BuildWatcher();
-
-    @ClassRule
-    @ConfiguredWithCode("jcasc-elastic-backend.yml")
-    public static JenkinsConfiguredWithCodeRule jenkinsRule = new JenkinsConfiguredWithCodeRule();
-
-    static OpenTelemetrySdkProvider openTelemetrySdkProvider;
-
-
-    @BeforeClass
-    public static void beforeClass() throws Exception {
-        System.out.println("beforeClass()");
-        System.out.println("Wait for jenkins to start...");
-        jenkinsRule.waitUntilNoActivity();
-        System.out.println("Jenkins started");
-
-        ExtensionList<OpenTelemetrySdkProvider> openTelemetrySdkProviders = jenkinsRule.getInstance().getExtensionList(OpenTelemetrySdkProvider.class);
-        verify(openTelemetrySdkProviders.size() == 1, "Number of openTelemetrySdkProviders: %s", openTelemetrySdkProviders.size());
-        openTelemetrySdkProvider = openTelemetrySdkProviders.get(0);
-
-        // verify(openTelemetrySdkProvider.openTelemetry == null, "OpenTelemetrySdkProvider has already been configured");
-        openTelemetrySdkProvider.initializeForTesting();
-
-        // openTelemetrySdkProvider.tracer.setDelegate(openTelemetrySdkProvider.openTelemetry.getTracer("jenkins"));
-    }
-
-    @Before
-    public void before() throws Exception {
-    }
-
-    @After
-    public void after() throws Exception {
-        jenkinsRule.waitUntilNoActivity();
-        ((InMemorySpanExporter) OpenTelemetrySdkProvider.TESTING_SPAN_EXPORTER).reset();
-        ((InMemoryMetricExporter) OpenTelemetrySdkProvider.TESTING_METRICS_EXPORTER).reset();
-    }
+public class JenkinsOtelPluginIntegrationTest extends BaseIntegrationTest {
 
     @Test
     public void testSimplePipeline() throws Exception {
@@ -135,26 +76,6 @@ public class JenkinsOtelPluginIntegrationTest {
         Collection<LongPointData> metricPoints = runCompletedCounterData.getLongSumData().getPoints();
         //MatcherAssert.assertThat(Iterables.getLast(metricPoints).getValue(), CoreMatchers.is(1L));
 
-
-    }
-
-    private void checkChainOfSpans(Tree<SpanDataWrapper> spanTree, String... expectedSpanNames) {
-        final List<String> expectedSpanNamesList = Arrays.asList(expectedSpanNames);
-        final Iterator<String> expectedSpanNamesIt = expectedSpanNamesList.iterator();
-        if (!expectedSpanNamesIt.hasNext()) {
-            Assert.fail("No element in the list of expected spans for " + Arrays.asList(expectedSpanNames));
-        }
-        final String leafSpanName = expectedSpanNamesIt.next();
-        Optional<Tree.Node<SpanDataWrapper>> actualNodeOptional = spanTree.breadthFirstSearchNodes(node -> Objects.equals(leafSpanName, node.getData().spanData.getName()));
-
-        MatcherAssert.assertThat("Expected leaf span '" + leafSpanName + "' in chain of span" + expectedSpanNamesList + " not found", actualNodeOptional.isPresent(), CoreMatchers.is(true));
-
-        while (expectedSpanNamesIt.hasNext()) {
-            String expectedSpanName = expectedSpanNamesIt.next();
-            actualNodeOptional = actualNodeOptional.get().getParent();
-            final String actualSpanName = actualNodeOptional.get().getData().spanData.getName();
-            MatcherAssert.assertThat("Expected span '" + expectedSpanName + "' in chain of span" + expectedSpanNamesList + " not found, actual is '" + actualSpanName + "'", actualSpanName, CoreMatchers.is(expectedSpanName));
-        }
     }
 
     @Test
@@ -272,34 +193,6 @@ public class JenkinsOtelPluginIntegrationTest {
         MatcherAssert.assertThat(spans.cardinality(), CoreMatchers.is(10L));
     }
 
-    private void dumpMetrics(Map<String, MetricData> exportedMetrics) {
-        System.out.println("Metrics: " + exportedMetrics.size());
-        System.out.println(exportedMetrics.values().stream().sorted(Comparator.comparing(MetricData::getName)).map(metric -> {
-            MetricDataType metricType = metric.getType();
-            String s = metric.getName() + "   " + metricType + " ";
-            switch (metricType) {
-                case LONG_SUM:
-                    s += metric.getLongSumData().getPoints().stream().map(point -> String.valueOf(point.getValue())).collect(Collectors.joining(", ")) + "";
-                    break;
-                case DOUBLE_SUM:
-                    s += metric.getDoubleSumData().getPoints().stream().map(point -> String.valueOf(point.getValue())).collect(Collectors.joining(", ")) + "";
-                    break;
-                case DOUBLE_GAUGE:
-                    s += metric.getDoubleGaugeData().getPoints().stream().map(point -> String.valueOf(point.getValue())).collect(Collectors.joining(", ")) + "";
-                    break;
-                case LONG_GAUGE:
-                    s += metric.getLongGaugeData().getPoints().stream().map(point -> String.valueOf(point.getValue())).collect(Collectors.joining(", ")) + "";
-                    break;
-                case SUMMARY:
-                    break;
-                default:
-
-            }
-            return s;
-        }).collect(Collectors.joining(" \n")));
-    }
-
-
     @Test
     public void testPipelineWithWrappingStep() throws Exception {
         String pipelineScript = "def xsh(cmd) {if (isUnix()) {sh cmd} else {bat cmd}};\n" +
@@ -377,83 +270,5 @@ public class JenkinsOtelPluginIntegrationTest {
         checkChainOfSpans(spans, "shell-3", "Parallel branch: parallelBranch3", "Stage: ze-parallel-stage", "Node", "Phase: Run");
         MatcherAssert.assertThat(spans.cardinality(), CoreMatchers.is(13L));
 
-    }
-
-	@Test
-	public void testMultibranchPipelineStep() throws Exception {
-		String pipelineScript = "def xsh(cmd) {if (isUnix()) {sh cmd} else {bat cmd}};\n" +
-				"node() {\n" +
-				"    stage('ze-parallel-stage') {\n" +
-				"        parallel parallelBranch1: {\n" +
-				"            xsh (label: 'shell-1', script: 'echo this-is-the-parallel-branch-1')\n" +
-				"        } ,parallelBranch2: {\n" +
-				"            xsh (label: 'shell-2', script: 'echo this-is-the-parallel-branch-2')\n" +
-				"        } ,parallelBranch3: {\n" +
-				"            xsh (label: 'shell-3', script: 'echo this-is-the-parallel-branch-3')\n" +
-				"        }\n" +
-				"    }\n" +
-				"}";
-		final Node node = jenkinsRule.createOnlineSlave();
-
-		WorkflowJob pipeline = jenkinsRule.createProject(WorkflowJob.class, "test-pipeline-with-parallel-step" + jobNameSuffix.incrementAndGet());
-		pipeline.setDefinition(new CpsFlowDefinition(pipelineScript, true));
-		WorkflowRun build = jenkinsRule.assertBuildStatus(Result.SUCCESS, pipeline.scheduleBuild2(0));
-
-		final Tree<SpanDataWrapper> spans = getGeneratedSpans();
-		checkChainOfSpans(spans, "shell-1", "Parallel branch: parallelBranch1", "Stage: ze-parallel-stage", "Node", "Phase: Run");
-		checkChainOfSpans(spans, "shell-2", "Parallel branch: parallelBranch2", "Stage: ze-parallel-stage", "Node", "Phase: Run");
-		checkChainOfSpans(spans, "shell-3", "Parallel branch: parallelBranch3", "Stage: ze-parallel-stage", "Node", "Phase: Run");
-		MatcherAssert.assertThat(spans.cardinality(), CoreMatchers.is(13L));
-
-	}
-
-    protected Tree<SpanDataWrapper> getGeneratedSpans() {
-
-        CompletableResultCode completableResultCode = this.openTelemetrySdkProvider.getOpenTelemetrySdk().getSdkTracerProvider().forceFlush();
-        completableResultCode.join(1, TimeUnit.SECONDS);
-        List<SpanData> spans = ((InMemorySpanExporter) OpenTelemetrySdkProvider.TESTING_SPAN_EXPORTER).getFinishedSpanItems();
-
-        final BiPredicate<Tree.Node<SpanDataWrapper>, Tree.Node<SpanDataWrapper>> parentChildMatcher = (spanDataNode1, spanDataNode2) -> {
-            final SpanData spanData1 = spanDataNode1.getData().spanData;
-            final SpanData spanData2 = spanDataNode2.getData().spanData;
-            return Objects.equals(spanData1.getSpanId(), spanData2.getParentSpanId());
-        };
-        final List<Tree<SpanDataWrapper>> trees = Tree.of(spans.stream().map(span -> new SpanDataWrapper(span)).collect(Collectors.toList()), parentChildMatcher);
-        System.out.println("## TREE VIEW OF SPANS ## ");
-        for (Tree<SpanDataWrapper> tree : trees) {
-            System.out.println(tree);
-        }
-
-        return trees.get(0);
-    }
-
-    @AfterClass
-    public static void afterClass() {
-        GlobalOpenTelemetry.resetForTest();
-    }
-
-    static class SpanDataWrapper {
-        final SpanData spanData;
-
-        public SpanDataWrapper(SpanData spanData) {
-            this.spanData = spanData;
-        }
-
-        @Override
-        public String toString() {
-            String result = spanData.getName();
-
-            final Attributes attributes = spanData.getAttributes();
-            if (attributes.get(JenkinsOtelSemanticAttributes.JENKINS_STEP_TYPE) != null) {
-                result += ", function: " + attributes.get(JenkinsOtelSemanticAttributes.JENKINS_STEP_TYPE);
-            }
-            if (attributes.get(JenkinsOtelSemanticAttributes.JENKINS_STEP_NAME) != null) {
-                result += ", name: " + attributes.get(JenkinsOtelSemanticAttributes.JENKINS_STEP_NAME);
-            }
-            if (attributes.get(JenkinsOtelSemanticAttributes.JENKINS_STEP_ID) != null) {
-                result += ", node.id: " + attributes.get(JenkinsOtelSemanticAttributes.JENKINS_STEP_ID);
-            }
-            return result;
-        }
     }
 }
